requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery[redis]==5.3.4
sentence-transformers>=2.2.2
huggingface-hub>=0.15.0
faiss-cpu==1.7.4
pypdf2==3.0.1
reportlab==4.0.7
httpx==0.25.2
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.24.3
redis>=4.5.2,<5.0.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0 